requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
numpy
pandas
Flask
scikit-learn
requests
Pillow
gunicorn == 20.0.4
torch==1.7.1+cpu
torchaudio==0.7.1
torchvision==0.8.2+cpu
pyrebase
